// frontend/src/lib/modules/knowledge-base/entities/types.ts
export interface RichText {
	type: 'text' | 'mention' | 'equation';
	text?: {
		content: string;
		link: string | null;
	};
	mention?: {
		type: 'user' | 'page' | 'date' | 'database';
		id: string;
	};
	equation?: {
		expression: string;
	};
	annotations: TextAnnotations;
	plain_text: string;
	href: string | null;
}

export interface TextAnnotations {
	bold: boolean;
	italic: boolean;
	strikethrough: boolean;
	underline: boolean;
	code: boolean;
	color: TextColor;
}

export type TextColor =
	| 'default'
	| 'gray'
	| 'brown'
	| 'orange'
	| 'yellow'
	| 'green'
	| 'blue'
	| 'purple'
	| 'pink'
	| 'red'
	| 'gray_background'
	| 'brown_background'
	| 'orange_background'
	| 'yellow_background'
	| 'green_background'
	| 'blue_background'
	| 'purple_background'
	| 'pink_background'
	| 'red_background';